lib/entity/pacman_test.py
from lib.utils.testing_utils import *
from lib.entity.players.pacman import Pacman
import pytest


class Test_movement:

    @pytest.fixture(autouse=True, scope="function")
    def pacman(self):
        return Pacman()

    @pytest.mark.parametrize("key,expected",
                             [(pygame.K_UP, pygame.Vector2(0, -1)), (pygame.K_DOWN, pygame.Vector2(0, 1)),
                              (pygame.K_LEFT, pygame.Vector2(-1, 0)), (pygame.K_RIGHT, pygame.Vector2(1, 0))])
    def test_handel_direction_change(self, pacman, map_with_walls, key, expected):
        pacman.handle_event(pygame.event.Event(pygame.KEYDOWN, {"key": key}))
        pacman.update(1e-10)
        # assert pacman.get_direction() == expected
        assert True
